move notification methods to page listener, rebase pr
